+++ src/cpx_io/cpx_system/cpx_base.py
@@ -5,7 +5,6 @@ from dataclasses import dataclass, fields
 from functools import wraps
 
 from pymodbus.client import ModbusTcpClient
-from pymodbus.exceptions import ModbusIOException
 from pymodbus.pdu.mei_message import ReadDeviceInformationRequest
 from cpx_io.utils.logging import Logging
 from cpx_io.utils.boollist import boollist_to_bytes, bytes_to_boollist
@@ -226,11 +225,11 @@ class CpxBase:
         # Convert to list of words
         reg = list(struct.unpack("<" + "H" * (len(data) // 2), data))
         # Write data
-        for offset, data in enumerate(reg):
-            return_value = self.client.write_register(register + offset, data)
+        for offset, d in enumerate(reg):
+            return_value = self.client.write_register(register + offset, d)
             retries = 3
             while return_value.isError():
-                return_value = self.client.write_register(register + offset, data)
+                return_value = self.client.write_register(register + offset, d)
                 if retries < 0:
                     break
                 retries -= 1